server/requirements.txt
pandas==2.3.3
fastapi==0.118.0
uvicorn==0.24.0
supabase==2.21.1
python-dotenv==1.0.0
python-multipart==0.0.18
openpyxl==3.1.2
google-generativeai>=0.3.0
